Moved QRCodeDecoder to qrcode module
